+++ data/setting.js
@@ -13,11 +13,25 @@ self.on('message', function(msg) {
       ctable = new CocoaTable(msg.data, ['name', 'format']);
       
       ctable._listener.onUpdated = function () {
-        postMessage({kind: "update", data: ctable.to_json()});
+        self.postMessage({kind: "update", data: ctable.to_json()});
       }
 
       ctable._listener.onPushButton = function (msg) {
-        postMessage({kind: msg});
+        self.postMessage({kind: msg});
+      }
+
+      {
+        var a = document.getElementById('link_recommended_settings');
+        a.addEventListener('click', function (ev) {
+          self.postMessage({kind: "tab_open", url:"http://firelink.ongaeshi.me/Recommended_Settings"});
+        }, true);
+      }
+
+      {
+        var a = document.getElementById('link_homepage');
+        a.addEventListener('click', function (ev) {
+          self.postMessage({kind: "tab_open", url:"http://firelink.ongaeshi.me/"});
+        }, true);
       }
     }
     break;